% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by-season.r, R/mlb-by-season.r
\name{active_players}
\alias{active_players}
\alias{mlb_active_players}
\title{List of active players}
\usage{
active_players(sport, season = "current")

mlb_active_players(season = "current")
}
\arguments{
\item{sport}{mlb | nfl | nba | etc..}

\item{season}{ex. 2017-regular | current}
}
\description{
List of active players
}
