% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-test.r
\name{capture_requests2}
\alias{capture_requests2}
\title{Capture requests to tests directory}
\usage{
capture_requests2(expr)
}
\arguments{
\item{expr}{R expresssion to execute}
}
\description{
Helper to build up mock fixtures for testing
}
\examples{
\dontrun{
capture_requests2(game_boxscore("nba", "20171027-BRO-NYK"))
}
}
\keyword{internal}
