% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by-date.r, R/mlb-by-date.r, R/nfl-by-date.r
\name{daily_dfs}
\alias{daily_dfs}
\alias{mlb_daily_dfs}
\alias{nfl_daily_dfs}
\title{Daily Fantasy Stats}
\usage{
daily_dfs(sport, date = Sys.Date(), season = "current")

mlb_daily_dfs(date = Sys.Date(), season = "current")

nfl_daily_dfs(date = Sys.Date(), season = "current")
}
\arguments{
\item{sport}{mlb | nfl | nba | etc..}

\item{date}{date, ex. as.Date("2017-05-08")}

\item{season}{ex. 2017-regular | current}
}
\description{
A list of players, along with their DFS salaries and actual fantasy points.
}
