% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by-date.r
\name{daily_game_schedule}
\alias{daily_game_schedule}
\title{Daily Game Schedule}
\usage{
daily_game_schedule(sport, date = Sys.Date(), season = "current")
}
\arguments{
\item{sport}{mlb | nfl | nba | etc..}

\item{date}{date, ex. as.Date("2017-05-08")}

\item{season}{ex. 2017-regular | current}
}
\description{
The daily game schedule
}
