% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by-date.r
\name{daily_player_stats}
\alias{daily_player_stats}
\title{Player Stats}
\usage{
daily_player_stats(sport, date = Sys.Date(), season = "current")
}
\arguments{
\item{sport}{mlb | nfl | nba | etc..}

\item{date}{date, ex. as.Date("2017-05-08")}

\item{season}{ex. 2017-regular | current}
}
\description{
A list of player stats totals for those players who particpated in games on a given day.
}
