% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by-game-parsers.r
\name{extract_lineup}
\alias{extract_lineup}
\title{Extract expected lineup}
\usage{
extract_lineup(x, team_id, is_mlb)
}
\arguments{
\item{x}{list of expected players in lineup}

\item{team_id}{id of team}
}
\description{
Extract expected lineup
}
