% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by-season.r
\name{full_game_schedule}
\alias{full_game_schedule}
\title{Full Game Schedule}
\usage{
full_game_schedule(sport, season = "current")
}
\arguments{
\item{sport}{mlb | nfl | nba | etc..}

\item{season}{ex. 2017-regular | current}
}
\description{
A list of all games to be played for the entire season.
}
