% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by-game.r
\name{game_boxscore}
\alias{game_boxscore}
\title{Game Boxscore}
\usage{
game_boxscore(sport, gameid, season = "current")
}
\arguments{
\item{sport}{mlb | nfl | nba | etc..}

\item{gameid}{string, "40265"}

\item{season}{string, ex. "2017-regular" | "2017-playoff"}
}
\description{
A boxscore of scoring plays and player stats, for a game.
}
