% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by-game.r, R/mlb-by-game.r, R/nfl-by-game.r
\name{game_pbp}
\alias{game_pbp}
\alias{mlb_game_pbp}
\alias{nfl_game_pbp}
\title{Game Play-by-Play}
\usage{
game_pbp(sport, gameid, season = "current")

mlb_game_pbp(gameid, season = "current")

nfl_game_pbp(gameid, season = "current")
}
\arguments{
\item{sport}{mlb | nfl | nba | etc..}

\item{gameid}{string, "40265"}

\item{season}{string, ex. "2017-regular" | "2017-playoff"}
}
\description{
All plays that have occurred for a game.
}
