% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by-game.r, R/mlb-by-game.r, R/nfl-by-game.r
\name{game_starting_lineup}
\alias{game_starting_lineup}
\alias{mlb_game_starting_lineup}
\alias{nfl_game_starting_lineup}
\title{Download starting lineups}
\usage{
game_starting_lineup(sport, gameid, season = "current")

mlb_game_starting_lineup(gameid, season = "current")

nfl_game_starting_lineup(gameid, season = "current")
}
\arguments{
\item{sport}{mlb | nfl | nba | etc..}

\item{gameid}{string, "40265"}

\item{season}{string, ex. "2017-regular" | "2017-playoff"}
}
\description{
Download starting lineups
}
