% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb-by-season.r
\name{mlb_active_players}
\alias{mlb_active_players}
\title{List of active players}
\usage{
mlb_active_players(season = "current")
}
\arguments{
\item{season}{ex. 2017-regular}
}
\description{
List of active players
}
