% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by-game-parsers.r
\name{mlb_batting_order}
\alias{mlb_batting_order}
\title{Parse mlb events}
\usage{
mlb_batting_order(id, position, team_id)
}
\arguments{
\item{event}{a nested json event}
}
\description{
Parse mlb events
}
\keyword{internal}
