% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb-by-season.r
\name{mlb_cumulative_player_stats}
\alias{mlb_cumulative_player_stats}
\title{Cumulative Player Stats}
\usage{
mlb_cumulative_player_stats(season = "current")
}
\arguments{
\item{season}{ex. 2017-regular}
}
\description{
A list of player stats totals for all roster players, summarized by their latest team.
}
