% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb.r
\name{mlb_daily_game_schedule}
\alias{mlb_daily_game_schedule}
\title{Daily MLB Game Schedule}
\usage{
mlb_daily_game_schedule(season = "current", date = format(Sys.Date(),
  "\%Y\%m\%d"))
}
\arguments{
\item{season}{ex. 2017-regular | current}

\item{date}{date format in Ymd such as 20170601}
}
\description{
The daily game schedule
}
