% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb.r
\name{mlb_daily_player_stats}
\alias{mlb_daily_player_stats}
\title{MLB Player Stats}
\usage{
mlb_daily_player_stats(date, season = "current")
}
\arguments{
\item{date}{string, ex. "20170508"}

\item{season}{string, ex. "2017-regular" | "2017-playoff"}
}
\description{
A list of player stats totals for those players who particpated in games on a given day.
}
