% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb.r
\name{mlb_full_game_schedule}
\alias{mlb_full_game_schedule}
\title{MLB Full Game Schedule}
\usage{
mlb_full_game_schedule(season)
}
\arguments{
\item{season}{string, ex. "2017-regular" | "2017-playoff"}
}
\description{
A list of all games to be played for the entire season.
}
