% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb-by-game.r
\name{mlb_game_boxscore}
\alias{mlb_game_boxscore}
\title{MLB Game Boxscore}
\usage{
mlb_game_boxscore(gameid, season = "current")
}
\arguments{
\item{gameid}{string, "40265"}

\item{season}{string, ex. "2017-regular" | "2017-playoff"}
}
\description{
A boxscore of scoring plays and player stats, for a game.
}
