% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb-by-game.r
\name{mlb_game_pbp}
\alias{mlb_game_pbp}
\title{MLB Game Play-by-Play}
\usage{
mlb_game_pbp(gameid, season = "current")
}
\arguments{
\item{gameid}{string, "40265"}

\item{season}{string, ex. "2017-regular" | "2017-playoff"}
}
\description{
All plays that have occurred for a game.
}
