% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb-by-game.r
\name{mlb_game_starting_lineup}
\alias{mlb_game_starting_lineup}
\title{Download starting mlb starting lineups}
\usage{
mlb_game_starting_lineup(gameid, season = "current")
}
\arguments{
\item{gameid}{the game id}

\item{season}{season such as 2017-regular | 2016-playoffs | current}
}
\description{
Download starting mlb starting lineups
}
