% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb-by-season.r
\name{mlb_latest_updates}
\alias{mlb_latest_updates}
\title{Latest Updates}
\usage{
mlb_latest_updates(season = "current")
}
\arguments{
\item{season}{ex. 2017-regular}
}
\description{
This feed lists the last updated date/time for each available feed.
}
