% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb-by-season.r
\name{mlb_player_gamelogs}
\alias{mlb_player_gamelogs}
\title{Player Game Logs}
\usage{
mlb_player_gamelogs(season = "current")
}
\arguments{
\item{season}{ex. 2017-regular}
}
\description{
Game logs for one or more players
}
