% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb-by-date.r
\name{mlb_scoreboard}
\alias{mlb_scoreboard}
\title{Scoreboard}
\usage{
mlb_scoreboard(date = format(Sys.Date(), "\%Y\%m\%d"), season = "current")
}
\arguments{
\item{date}{date format in Ymd such as 20170601}

\item{season}{ex. 2017-regular | current}
}
\description{
Scores and status for all games on a given day.
}
