% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by-game.r
\name{msf_by_game}
\alias{msf_by_game}
\title{MySportsFeeds API calls by game}
\usage{
msf_by_game(req, sport, gameid, season)
}
\arguments{
\item{req}{data to request}

\item{sport}{mlb | nfl | nba | etc..}

\item{gameid}{string, "40265"}

\item{season}{string, ex. "2017-regular" | "2017-playoff"}
}
\description{
MySportsFeeds API calls by game
}
