% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by-team.r
\name{msf_by_team}
\alias{msf_by_team}
\alias{team_gamelogs}
\alias{player_gamelogs}
\title{MySportsFeeds API calls by team}
\usage{
msf_by_team(feed, sport, team, season = "current", delay = 1)

team_gamelogs(sport, team, ...)

player_gamelogs(sport, team, ...)
}
\arguments{
\item{feed}{feed to request}

\item{sport}{mlb | nfl | nba | etc..}

\item{team}{team id or abbreviation}

\item{season}{ex. 2017-regular | current}

\item{delay}{number of seconds between queries}

\item{...}{additional parameters passed to \code{msf_by_team} such as season}
}
\description{
If you provide multiple teams it will execute with a delay between each query.
This is to ensure you obey limit rates but you are free to lower the parameter if you don't
think your queries will reach the throttle limits.
}
\section{Functions}{
\itemize{
\item \code{team_gamelogs}: Game logs for one or more teams

\item \code{player_gamelogs}: Game logs for one or more players
}}

\examples{
\dontrun{
resp <- team_gamelogs("nba", c("bos", "atl"))
}
}
