% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba.r
\name{nba_daily_dfs}
\alias{nba_daily_dfs}
\title{NBA Daily Fantasy Stats}
\usage{
nba_daily_dfs(season, date)
}
\arguments{
\item{season}{string, ex. "2016-2017-regular"}

\item{date}{string, ex. "20161104"}
}
\description{
A list of players, along with their DFS salaries and actual fantasy points.
}
