% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba.r
\name{nba_daily_player_stats}
\alias{nba_daily_player_stats}
\title{NBA Player Stats}
\usage{
nba_daily_player_stats(season, date)
}
\arguments{
\item{season}{string, ex. "2016-2017-regular"}

\item{date}{string, ex. "20161104"}
}
\description{
A list of player stats totals for those players who particpated in games on a given day.
}
