% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba.r
\name{nba_full_game_schedule}
\alias{nba_full_game_schedule}
\title{NBA Full Game Schedule}
\usage{
nba_full_game_schedule(season)
}
\arguments{
\item{season}{string, ex. "2016-2017-regular" | "2017-playoff"}
}
\description{
A list of all games to be played for the entire season.
}
