% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba.r
\name{nba_game_boxscore}
\alias{nba_game_boxscore}
\title{NBA Game Boxscore}
\usage{
nba_game_boxscore(season, gameid)
}
\arguments{
\item{season}{string, ex. "2016-2017-regular"}

\item{gameid}{string, "33941"}
}
\description{
A boxscore of scoring plays and player stats, for a game.
}
