% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by-season-parsers.r
\name{parse_active_players}
\alias{parse_active_players}
\title{Parse active players}
\usage{
parse_active_players(json)
}
\arguments{
\item{json}{content from response}
}
\description{
Parse active players
}
\examples{
\dontrun{
resp <- mysportsfeeds::mlb_active_players()
parse_active_players(resp$content)
}
}
