% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by-game-parsers.r
\name{parse_boxscore}
\alias{parse_boxscore}
\title{Parse box scores}
\usage{
parse_boxscore(json)
}
\arguments{
\item{json}{content from response}
}
\description{
Parse box scores
}
\examples{
\dontrun{
resp <- mysportsfeeds::nfl_game_boxscore("30904", "2016-2017-regular")
parse_boxscore(resp$content)
}
}
