% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by-season-parsers.r
\name{parse_cumulative_stats}
\alias{parse_cumulative_stats}
\title{Parse cumulative stats}
\usage{
parse_cumulative_stats(json)
}
\arguments{
\item{json}{content from response}
}
\description{
Parse cumulative stats
}
\examples{
\dontrun{
resp <- mlb_cumulative_player_stats()
stats <- parse_cumulative_stats(resp$content)
}
}
