% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by-date-parsers.r
\name{parse_current_season}
\alias{parse_current_season}
\title{Parse Current Season}
\usage{
parse_current_season(json)
}
\arguments{
\item{json}{list of json content}
}
\description{
Parse Current Season
}
\examples{
\dontrun{
resp <- current_season("nba")
parse_current_season(resp$content)
}
}
