% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by-date-parsers.r
\name{parse_daily_dfs}
\alias{parse_daily_dfs}
\title{Parse fantasy salaries for today's games}
\usage{
parse_daily_dfs(json, site = c("draftkings", "fanduel"))
}
\arguments{
\item{json}{content from response}

\item{site}{draftkings or fanduel}
}
\description{
Parse fantasy salaries for today's games
}
\examples{
\dontrun{
resp <- mysportsfeeds::mlb_daily_dfs()
salaries <- parse_daily_dfs(resp$json)
}
}
