% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by-date-parsers.r
\name{parse_game_schedule}
\alias{parse_game_schedule}
\alias{parse_daily_game_schedule}
\title{Parse game schedules}
\usage{
parse_game_schedule(json, nm)

parse_daily_game_schedule(json)
}
\arguments{
\item{json}{content from response}

\item{nm}{"dailygameschedule" or "fullgameschedule"}
}
\description{
Parse game schedules
}
\section{Functions}{
\itemize{
\item \code{parse_daily_game_schedule}: Parse daily game schedule
}}

\examples{
\dontrun{
resp <- daily_game_schedule("nfl")
parse_daily_game_schedule(resp$json)
}
}
