% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by-season-parsers.r
\name{parse_latest_updates}
\alias{parse_latest_updates}
\title{Parse latest updates}
\usage{
parse_latest_updates(json)
}
\arguments{
\item{json}{content from response}
}
\description{
Parse latest updates into a tidy data frame.
This function is pretty slow right now. Needs to be sped up in the future.
}
\examples{
\dontrun{
resp <- latest_updates("nba")
updates <- parse_latest_updates(resp$content)
}
}
