% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by-game-parsers.r
\name{parse_mlb_event}
\alias{parse_mlb_event}
\title{Parse mlb events}
\usage{
parse_mlb_event(event)
}
\arguments{
\item{event}{a nested json event}
}
\description{
Parse mlb events
}
\keyword{internal}
