% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by-date-parsers.r
\name{parse_roster_players}
\alias{parse_roster_players}
\title{Parse roster players}
\usage{
parse_roster_players(json)
}
\arguments{
\item{json}{content from response}
}
\description{
Parse roster players
}
\examples{
\dontrun{
resp <- roster_players("mlb", season = "2017-regular")
parse_roster_players(resp$content)
}
}
