% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by-game-parsers.r
\name{parse_starting_lineup}
\alias{parse_starting_lineup}
\title{Parse starting lineup for a game}
\usage{
parse_starting_lineup(json, type = c("actual", "expected"))
}
\arguments{
\item{json}{content from response}

\item{type}{actual or expected lineup}
}
\description{
Parse starting lineup for a game
}
\examples{
\dontrun{
resp <- game_starting_lineup("nhl", "20171014-BUF-LAK", season = "2017-2018-regular")
resp <- game_starting_lineup("mlb", "20170822-COL-KC", season = "2017-regular")
parse_starting_lineup(resp$content, "actual")

}
}
