% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-parsers.r
\name{parse_stats}
\alias{parse_stats}
\title{Parse stats data from json}
\usage{
parse_stats(json)
}
\arguments{
\item{json}{list of json data}
}
\description{
Across most MySportsFeeds data there is usually a common format for stats data.
This usually includes a number of numeric fields. This function
simply pulls out these fields for you so you don't have to copy code.
}
\keyword{internal}
