% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-parsers.r
\name{parse_teams}
\alias{parse_teams}
\title{Parse team data from json}
\usage{
parse_teams(json)
}
\arguments{
\item{json}{list of json data}
}
\description{
Across most MySportsFeeds data there is usually a common format for team data.
This usually includes the team id and abbreviation. This function
simply pulls out these fields for you so you don't have to copy code.
}
\keyword{internal}
