% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by-season.r
\name{player_injuries}
\alias{player_injuries}
\title{Player Injuries}
\usage{
player_injuries(sport, season = "current")
}
\arguments{
\item{sport}{mlb | nfl | nba | etc..}

\item{season}{ex. 2017-regular | current}
}
\description{
A list of all currently injured players.
}
