% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-test.r
\name{with_mock_test}
\alias{with_mock_test}
\title{Mock API calls with fake credentials}
\usage{
with_mock_test(expr, user = "fakeuser", password = "fakepass")
}
\arguments{
\item{expr}{R expression}

\item{user, password}{strings}
}
\description{
Mock API calls with fake credentials
}
\keyword{internal}
