% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.r
\name{write_msf}
\alias{write_msf}
\title{Write MySportsFeeds JSON to disk}
\usage{
write_msf(resp, path)
}
\arguments{
\item{resp}{response or list of reponses}

\item{path}{path to directory}
}
\description{
Files will be named by the name variable in the object. Ex. resp$name
}
