% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by-season.r
\name{cumulative_player_stats}
\alias{cumulative_player_stats}
\title{Cumulative Player Stats}
\usage{
cumulative_player_stats(sport, season = "current")
}
\arguments{
\item{sport}{mlb | nfl | nba | etc..}

\item{season}{ex. 2017-regular | current}
}
\description{
A list of player stats totals for all roster players, summarized by their latest team.
}
