% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by-season.r, R/mlb-by-season.r,
%   R/nba-by-season.r, R/nfl-by-season.r
\name{full_game_schedule}
\alias{full_game_schedule}
\alias{mlb_full_game_schedule}
\alias{nba_full_game_schedule}
\alias{nfl_full_game_schedule}
\title{Full Game Schedule}
\usage{
full_game_schedule(sport, season = "current")

mlb_full_game_schedule(season = "current")

nba_full_game_schedule(season = "current")

nfl_full_game_schedule(season = "current")
}
\arguments{
\item{sport}{mlb | nfl | nba | etc..}

\item{season}{ex. 2017-regular | current}
}
\description{
A list of all games to be played for the entire season.
}
