% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by-season.r
\name{latest_updates}
\alias{latest_updates}
\title{Latest Updates}
\usage{
latest_updates(sport, season = "current")
}
\arguments{
\item{sport}{mlb | nfl | nba | etc..}

\item{season}{ex. 2017-regular | current}
}
\description{
This feed lists the last updated date/time for each available feed.
}
