% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb.r
\name{mlb_daily_dfs}
\alias{mlb_daily_dfs}
\title{MLB Daily Fantasy Stats}
\usage{
mlb_daily_dfs(season, date)
}
\arguments{
\item{season}{string, ex. "2017-regular" | "2017-playoff"}

\item{date}{string, ex. "20170508"}
}
\description{
A list of players, along with their DFS salaries and actual fantasy points.
}
