% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb-by-date.r
\name{mlb_daily_game_schedule}
\alias{mlb_daily_game_schedule}
\title{Daily MLB Game Schedule}
\usage{
mlb_daily_game_schedule(date = format(Sys.Date(), "\%Y\%m\%d"),
  season = "current")
}
\arguments{
\item{date}{date format in Ymd such as 20170601}

\item{season}{ex. 2017-regular | current}
}
\description{
The daily game schedule
}
