% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by-date.r
\name{msf_by_date}
\alias{msf_by_date}
\title{MySportsFeeds API calls by date}
\usage{
msf_by_date(req, sport, date, season)
}
\arguments{
\item{req}{data to request}

\item{sport}{mlb | nfl | nba | etc..}

\item{date}{date, ex. as.Date("2017-05-08")}

\item{season}{ex. 2017-regular | current}
}
\description{
MySportsFeeds API calls by date
}
