% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by-season.r
\name{msf_by_season}
\alias{msf_by_season}
\title{MySportsFeeds API calls by season}
\usage{
msf_by_season(req, sport, season)
}
\arguments{
\item{req}{data to request}

\item{sport}{mlb | nfl | nba | etc..}

\item{season}{ex. 2017-regular | current}
}
\description{
MySportsFeeds API calls by season
}
