% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by-game-parsers.r
\name{parse_boxscore}
\alias{parse_boxscore}
\title{Parse box scores}
\usage{
parse_boxscore(json)
}
\arguments{
\item{json}{content from response}
}
\description{
Parse box scores
}
\examples{
\dontrun{
resp <- game_boxscore("nfl", "20170917-ARI-IND", season = "2017-2018-regular")
resp <- game_boxscore("nhl", "20171114-BUF-PIT", season = "2017-2018-regular")
parse_boxscore(resp$content)
}
}
