% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by-game-parsers.r
\name{parse_game_pbp}
\alias{parse_game_pbp}
\title{Parse Play by Play Data}
\usage{
parse_game_pbp(json, sport = c(NA, "nba", "nhl", "nfl", "mlb"))
}
\arguments{
\item{json}{list of data}

\item{sport}{sport}
}
\description{
Parse Play by Play Data
}
\examples{
\dontrun{
resp <- game_pbp("nhl", "20161215-FLO-WPJ", season = "2016-2017-regular")
parse_game_pbp(resp$content, "nhl")
}
}
