% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by-date-parsers.r
\name{parse_game_schedule}
\alias{parse_game_schedule}
\title{Parse json of game schedule}
\usage{
parse_game_schedule(json, nm = "dailygameschedule")
}
\arguments{
\item{json}{content from response}
}
\description{
Parse json of game schedule
}
\examples{
\dontrun{
resp <- mysportsfeeds::mlb_daily_game_schedule()
games <- parse_daily_game_schedule(resp$json)
}
}
