% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by-game-parsers.r
\name{parse_starting_lineup}
\alias{parse_starting_lineup}
\title{Parse starting lineup for a game}
\usage{
parse_starting_lineup(json, type = c("expected", "actual"), is_mlb = FALSE)
}
\arguments{
\item{json}{content from response}

\item{type}{actual or expected lineup}

\item{is_mlb}{whether to parse batting order or not}
}
\description{
Parse starting lineup for a game
}
\examples{
\dontrun{
resp <- mysportsfeeds::mlb_game_starting_lineup("38494")
starters <- mlb_starters(resp$json)
}
}
