% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by-season.r, R/mlb-by-season.r,
%   R/nfl-by-season.r
\name{player_gamelogs}
\alias{player_gamelogs}
\alias{mlb_player_gamelogs}
\alias{nfl_player_gamelogs}
\title{Player Game Logs}
\usage{
player_gamelogs(sport, season = "current")

mlb_player_gamelogs(season = "current")

nfl_player_gamelogs(season = "current")
}
\arguments{
\item{sport}{mlb | nfl | nba | etc..}

\item{season}{ex. 2017-regular | current}
}
\description{
Game logs for one or more players
}
