% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by-date.r, R/mlb-by-date.r, R/nfl-by-date.r
\name{scoreboard}
\alias{scoreboard}
\alias{mlb_scoreboard}
\alias{nfl_scoreboard}
\title{Scoreboard}
\usage{
scoreboard(sport, date = Sys.Date(), season = "current")

mlb_scoreboard(date = Sys.Date(), season = "current")

nfl_scoreboard(date = Sys.Date(), season = "current")
}
\arguments{
\item{sport}{mlb | nfl | nba | etc..}

\item{date}{date, ex. as.Date("2017-05-08")}

\item{season}{ex. 2017-regular | current}
}
\description{
Scores and status for all games on a given day.
}
